Break transform module into several files
