+++ src/repl/mod.rs
@@ -11,8 +11,9 @@ use crate::error::{Error, EvalError, IOError, ParseError, ParseErrors, ReplError
 use crate::identifier::Ident;
 use crate::parser::{grammar, lexer, ExtendedTerm};
 use crate::term::{RichTerm, Term};
+use crate::transform::import_resolution;
 use crate::types::Types;
-use crate::{eval, transformations, typecheck};
+use crate::{eval, transform, typecheck};
 use codespan::FileId;
 use simple_counter::*;
 use std::ffi::{OsStr, OsString};
@@ -121,7 +122,7 @@ impl ReplImpl {
             // Because we don't use the cache for input, we have to perform recursive import
             // resolution/typechecking/transformation by oursleves.
             ExtendedTerm::RichTerm(t) => {
-                let (t, pending) = transformations::resolve_imports(t, &mut self.cache)?;
+                let (t, pending) = import_resolution::resolve_imports(t, &mut self.cache)?;
                 for id in &pending {
                     self.cache.resolve_imports(*id).unwrap();
                 }
@@ -135,7 +136,7 @@ impl ReplImpl {
                         })?;
                 }
 
-                let t = transformations::transform(t);
+                let t = transform::transform(t);
                 for id in &pending {
                     self.cache
                         .transform(*id)
@@ -145,7 +146,7 @@ impl ReplImpl {
                 Ok(eval_function(t, &self.env.eval_env, &mut self.cache)?.into())
             }
             ExtendedTerm::ToplevelLet(id, t) => {
-                let (t, pending) = transformations::resolve_imports(t, &mut self.cache)?;
+                let (t, pending) = import_resolution::resolve_imports(t, &mut self.cache)?;
                 for id in &pending {
                     self.cache.resolve_imports(*id).unwrap();
                 }
@@ -160,7 +161,7 @@ impl ReplImpl {
                         })?;
                 }
 
-                let t = transformations::transform(t);
+                let t = transform::transform(t);
                 for id in &pending {
                     self.cache
                         .transform(*id)
@@ -207,7 +208,7 @@ impl Repl for ReplImpl {
         })?;
 
         let term = self.cache.get_owned(file_id).unwrap();
-        let (term, pending) = transformations::resolve_imports(term, &mut self.cache)?;
+        let (term, pending) = import_resolution::resolve_imports(term, &mut self.cache)?;
         for id in &pending {
             self.cache.resolve_imports(*id).unwrap();
         }
@@ -221,7 +222,7 @@ impl Repl for ReplImpl {
         let file_id = self.cache.add_tmp("<repl-typecheck>", String::from(exp));
         // We ignore non fatal errors while type checking.
         let (term, _) = self.cache.parse_nocache(file_id)?;
-        let (term, pending) = transformations::resolve_imports(term, &mut self.cache)?;
+        let (term, pending) = import_resolution::resolve_imports(term, &mut self.cache)?;
         for id in &pending {
             self.cache.resolve_imports(*id).unwrap();
         }
